// app/back/src/domain/models/click.ts
export interface ClickModel {
  id: number
  id_click: string
  id_campaign: number
  id_campaign_taboola: string
  id_ads_taboola: string
  id_site: string
  site: string
  thumbnail: string
  title: string
  step_1: number
  step_2: number
  step_3: number
  checkout: number
  revenue: number
  payment_type: string
  id_order: string
  created_at: Date
}